requirements.txt
DateTime==4.3
pandas==1.3.3
DateTime==4.3
python-dateutil==2.8.2
streamlit==1.22.0
mysqlclient